+++ client/src/components/shared/Nav/Nav.jsx
@@ -6,7 +6,6 @@ import Search from '../../Search/Search.jsx'
 const Nav = (props) => {
 
   const { allProducts, setAllProducts } = props;
-  console.log(allProducts)
 
   return (
     <nav>
